slight include file ordering tweak
2003-08-13  Manish Singh  <yosh@gimp.org>

        * pygimp/Makefile.am: slight include file ordering tweak

        * plug-ins/pygimp/plug-ins/colorhtml.py: add a sanity check

        * plug-ins/pygimp/plug-ins/sphere.py
        * plug-ins/pygimp/plug-ins/whirlpinch.py: minor cleanups for
        Python 2.3

        * plug-ins/webbrowser/web-browser.scm: change script to reflect
        new pdb name of the web-browser.

+++ plug-ins/pygimp/plug-ins/whirlpinch.py
@@ -193,7 +193,7 @@ def bilinear(x, y, values):
 	y = y % 1.0
 	m0 = values[0] + x * (values[1] - values[0])
 	m1 = values[2] + x * (values[3] - values[2])
-	return chr(m0 + y * (m1 - m0))
+	return chr(int(m0 + y * (m1 - m0)))
 
 
 register(
